+++ tests/testthat/test-lgumbel.R
@@ -14,16 +14,23 @@
 
 context("lgumbel")
 
-test_that("rlgumbel", {
-  set.seed(99)
-  r <- rlgumbel(100000, location = 100, scale = 3)
-  expect_identical(length(r), 100000L)
-  expect_equal(mean(log(r)), 3 * 0.57721 + 100, tolerance = 0.001)
+test_that("dlgumbel", {
+  expect_equal(dlgumbel(exp(3), log(3), log(1)), 0.01831564, tolerance = 0.0000001)
+  expect_equal(dlgumbel(exp(4), log(3), log(1)), 0.004664011, tolerance = 0.0000001)
+})
+
+test_that("fit lgumbel", {
+  dist <- ssdtools:::ssd_fit_dist(ssdtools::boron_data, dist = "lgumbel")
+
+  expect_true(is.fitdist(dist))
+  expect_equal(coef(dist),
+  c(locationlog = 0.653681627422279, scalelog = 0.208983173310441
+))
 })
 
 test_that("pqlgumbel", {
-  expect_equal(log(qlgumbel(c(0.2, 0.5, 0.9), 3, 1)), c(2.52411500467289, 3.36651292058166, 5.25036732731245))
-  expect_equal(log(qlgumbel(c(0.2, 0.5, 0.9), 3, 1,
+  expect_equal(log(qlgumbel(c(0.2, 0.5, 0.9), log(3), log(1))), c(2.52411500467289, 3.36651292058166, 5.25036732731245))
+  expect_equal(log(qlgumbel(c(0.2, 0.5, 0.9), log(3), log(1),
     lower.tail = FALSE
   )), c(4.49993998675952, 3.36651292058166, 2.16596755475204))
 
@@ -32,12 +39,14 @@ test_that("pqlgumbel", {
     log(qlgumbel(exp(-1)))
   )
 
-  expect_equal(plgumbel(exp(3), 3, 1), 0.3678794, tolerance = 0.0000001)
-  expect_equal(plgumbel(exp(4), 3, 1), 0.6922006, tolerance = 0.0000001)
+  expect_equal(plgumbel(exp(3), log(3), log(1)), 0.3678794, tolerance = 0.0000001)
+  expect_equal(plgumbel(exp(4), log(3), log(1)), 0.6922006, tolerance = 0.0000001)
   expect_identical(plgumbel(qlgumbel(0.5, 3, 1), 3, 1), 0.5)
 })
 
-test_that("dlgumbel", {
-  expect_equal(dlgumbel(exp(3), 3, 1), 0.01831564, tolerance = 0.0000001)
-  expect_equal(dlgumbel(exp(4), 3, 1), 0.004664011, tolerance = 0.0000001)
+test_that("rlgumbel", {
+  set.seed(99)
+  r <- rlgumbel(100000, location = log(100), scale = log(3))
+  expect_identical(length(r), 100000L)
+  expect_equal(mean(log(r)), 3 * 0.57721 + 100, tolerance = 0.001)
 })